refactor: centralize tile color mapping
