+++ src/Board.jsx
@@ -11,20 +11,7 @@ import Monster from './entities/Monster';
 import { GameContext } from './GameContext';
 import loadWorld from './maps/loadWorld';
 import stepSfx from './assets/sounds/step.mp3';
-
-
-const tileColors = {
-  floor: '#8bc34a',  // 초원
-  wall: '#9e9e9e',   // 벽
-  water: '#42a5f5',  // 바다
-  special: '#ffd700', // 특별 지형
-  desert: '#e0c070',
-  ice: '#a0e0ff',
-  forest: '#228b22',
-  ocean: '#1e90ff',
-  jungle: '#006400',
-  grassland: '#7cfc00'
-};
+import tileColors from './utils/tileColors';
 
 const BOARD_SIZE = 7;
 const CENTER = Math.floor(BOARD_SIZE / 2);
