+++ src/utils/tileColors.js
@@ -1,0 +1,14 @@
+const tileColors = {
+  floor: '#8bc34a', // 초원
+  wall: '#9e9e9e', // 벽
+  water: '#42a5f5', // 바다
+  special: '#ffd700', // 특별 지형
+  desert: '#e0c070',
+  ice: '#a0e0ff',
+  forest: '#228b22',
+  ocean: '#1e90ff',
+  jungle: '#006400',
+  grassland: '#7cfc00',
+};
+
+export default tileColors;
